rename common to bpf_uapi
The previous name was too generic and did not clearly communicate the
purpose of the directory.
Rename it to `bpf_uapi` to emphasize that it contains the shared UAPI
definitions used by both userspace components and BPF programs.
