+++ bpf_uapi/action_rules.h
@@ -1,7 +1,7 @@
 /**
  *	Tempesta xFW action structures common for eBPF and manager
  *
- * See the rules for shared struct declarations in bpf_uapi.h.
+ * See the rules for shared struct declarations in types.h.
  *
  * SPDX-FileCopyrightText: © 2026 Tempesta Technologies, Inc.
  * SPDX-License-Identifier: GPL-2.0-or-later
