+++ bpf_uapi/config.h
@@ -1,7 +1,7 @@
 /**
  *	Tempesta xFW configuration structures
  *
- * See the rules for shared struct declarations in bpf_uapi.h.
+ * See the rules for shared struct declarations in types.h.
  *
  * SPDX-FileCopyrightText: © 2026 Tempesta Technologies, Inc.
  * SPDX-License-Identifier: GPL-2.0-or-later
@@ -13,7 +13,7 @@
 #include "tcp_control_bits.h"
 
 #include "bpf_bitset.h"
-#include "bpf_uapi_action_rules.h"
+#include "action_rules.h"
 
 #include "limits.h"
