+++ bpf_uapi/stats.h
@@ -1,7 +1,7 @@
 /**
  *	Tempesta xFW common statistic structures for eBPF and user-space
  *
- * See the rules for the declarations in bpf_uapi.h.
+ * See the rules for the declarations in types.h.
  *
  * SPDX-FileCopyrightText: © 2026 Tempesta Technologies, Inc.
  * SPDX-License-Identifier: GPL-2.0-or-later
