+++ manager/xfw/map_key_hash.hh
@@ -6,7 +6,7 @@
  */
 #pragma once
 #include <string.h>
-#include "../../common/bpf_uapi_proto_keys.h"
+#include "../../bpf_uapi/proto_keys.h"
 
 template<typename T>
 inline void
